+++ src/request/transaction.spec.ts
@@ -266,9 +266,7 @@ describe('Transaction()', () => {
 
       expect(test.toString()).toContain(deviceString);
 
-      expect(test.toString()).toContain(
-        '"credit_card":{"last_4_digits":"1234"}'
-      );
+      expect(test.toString()).toContain('"credit_card":{"last_digits":"1234"}');
     });
 
     it('it handles optional order field', () => {
@@ -317,7 +315,7 @@ describe('Transaction()', () => {
   });
 
   describe('key casing conversion', () => {
-    describe('`creditCard.last4digits` => `creditCard.last_4_digits`', () => {
+    describe('`creditCard.lastDigits/last4digits` => `creditCard.lastDigits`', () => {
       test('typed value is mapped', () => {
         const test = JSON.parse(
           new Transaction({
@@ -328,7 +326,7 @@ describe('Transaction()', () => {
           }).toString()
         );
 
-        expect(test.credit_card).toHaveProperty('last_4_digits', '1234');
+        expect(test.credit_card).toHaveProperty('last_digits', '1234');
       });
 
       test('null value is mapped', () => {
@@ -343,7 +341,35 @@ describe('Transaction()', () => {
           }).toString()
         );
 
-        expect(test.credit_card).toHaveProperty('last_4_digits', null);
+        expect(test.credit_card).toHaveProperty('last_digits', null);
+      });
+
+      test('typed value is mapped', () => {
+        const test = JSON.parse(
+          new Transaction({
+            device: new Device({
+              ipAddress: '1.1.1.1',
+            }),
+            creditCard: new CreditCard({ lastDigits: '1234' }),
+          }).toString()
+        );
+
+        expect(test.credit_card).toHaveProperty('last_digits', '1234');
+      });
+
+      test('null value is mapped', () => {
+        const test = JSON.parse(
+          new Transaction({
+            device: new Device({
+              ipAddress: '1.1.1.1',
+            }),
+            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+            // @ts-ignore explicit null
+            creditCard: new CreditCard({ lastDigits: null }),
+          }).toString()
+        );
+
+        expect(test.credit_card).toHaveProperty('last_digits', null);
       });
     });
 
@@ -413,6 +439,65 @@ describe('Transaction()', () => {
 
         expect(test.shipping).toHaveProperty('address_2', null);
       });
+    });
+  });
+
+  describe('6 or 8 digit iins and 2 or 4 digit lastDigits', () => {
+    it('it handles 8 digit iins with 2 digit lastDigits', () => {
+      const test = new Transaction({
+        creditCard: new CreditCard({
+          issuerIdNumber: '12345678',
+          lastDigits: '12',
+        }),
+        device: new Device({
+          ipAddress: '1.1.1.1',
+          sessionAge: 100,
+        }),
+      });
+
+      expect(isJSON(test.toString())).toBe(true);
+
+      expect(test.toString()).toContain(
+        '"credit_card":{"issuer_id_number":"12345678","last_digits":"12"}'
+      );
+    });
+
+    it('it handles 8 digit iins with 4 digit lastDigits', () => {
+      const test = new Transaction({
+        creditCard: new CreditCard({
+          issuerIdNumber: '12345678',
+          lastDigits: '1234',
+        }),
+        device: new Device({
+          ipAddress: '1.1.1.1',
+          sessionAge: 100,
+        }),
+      });
+
+      expect(isJSON(test.toString())).toBe(true);
+
+      expect(test.toString()).toContain(
+        '"credit_card":{"issuer_id_number":"12345678","last_digits":"1234"}'
+      );
+    });
+
+    it('it handles 6 digit iins with 2 digit lastDigits', () => {
+      const test = new Transaction({
+        creditCard: new CreditCard({
+          issuerIdNumber: '123456',
+          lastDigits: '12',
+        }),
+        device: new Device({
+          ipAddress: '1.1.1.1',
+          sessionAge: 100,
+        }),
+      });
+
+      expect(isJSON(test.toString())).toBe(true);
+
+      expect(test.toString()).toContain(
+        '"credit_card":{"issuer_id_number":"123456","last_digits":"12"}'
+      );
     });
   });
 });